backend/requirements.txt
# FastAPI & Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# AI & ML
openai==1.3.7
anthropic==0.7.7
httpx==0.25.2
openai-whisper==20231117
torch==2.1.1
torchaudio==2.1.1

# TTS (Text-to-Speech)
TTS==0.21.3
pydub==0.25.1
soundfile==0.12.1
librosa==0.10.1

# Database
supabase==2.3.0
python-jose[cryptography]==3.3.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
aiofiles==23.2.1
jinja2==3.1.2

# Audio Processing
numpy==1.26.2
scipy==1.11.4
webrtcvad==2.0.10

# Encryption & Security
cryptography==41.0.7
bcrypt==4.1.2

# Sentiment Analysis
transformers==4.36.2
sentencepiece==0.1.99

# Testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
